WIP: T01 -- Split tests/integration/test_websocket_service.py into a websocket/ subpackage
